Move Sitebar to components
